# doc/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AriesHub API",
    "version": "0.7.0",
    "description": "Public content catalog, Redis-backed cookie identity, flat paged discussions with comment likes, administrator content and moderation API. Credit balance lives on users and pricing on publications; the credit ledger and unlock tables exist but the credit application API is not implemented yet. Private S3-compatible file uploads and owner-only temporary downloads. All JSON responses use Result {code,msg,data,traceId}; success code is SUCCESS, failures retain semantic HTTP status and business codes. Empty successes use HTTP 200 with data:null; throttling includes Retry-After and data.retryAfterSeconds."
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1"
    }
  ],
  "paths": {
...
    "/admin/discussions/comments/{id}/hide": {
      "post": {
        "operationId": "hideComment",
        "summary": "管理员隐藏评论；整条不再对外可见",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success with an empty result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "code",
                    "msg",
                    "data",
                    "traceId"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "SUCCESS"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "type": "null"
                    },
                    "traceId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "401": {
            "description": "需要登录",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "403": {
            "description": "需要管理员角色",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "404": {
            "description": "评论不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "500": {
            "description": "服务端异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      }
    },
    "/admin/discussions/threads/lock": {
      "post": {
        "operationId": "lockThread",
        "summary": "管理员锁定讨论；线程仍可读，但不再接受评论与回复",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LockThreadRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success with an empty result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "code",
                    "msg",
                    "data",
                    "traceId"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "SUCCESS"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "type": "null"
                    },
                    "traceId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "400": {
            "description": "目标参数不合法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "401": {
            "description": "需要登录",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "403": {
            "description": "需要管理员角色",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "404": {
            "description": "目标不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "500": {
            "description": "服务端异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "msg",
          "data",
          "traceId"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "retryAfterSeconds"
                ],
                "properties": {
                  "retryAfterSeconds": {
                    "type": "integer",
                    "minimum": 1
                  }
                }
              }
            ]
          },
          "traceId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "LockThreadRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "targetType",
          "targetKey"
        ],
        "properties": {
          "targetType": {
            "type": "string",
            "pattern": "^[A-Z][A-Z0-9_]{1,39}$",
            "example": "PUBLICATION"
          },
          "targetKey": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120,
            "example": "ai-ppt-outline"
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "arieshub_token"
      }
    }
  }
}